Amend to a more suitable type.

+++ src/vesper/lib/types.h
@@ -17,7 +17,7 @@
 typedef uint32_t size_t;
 typedef uint32_t addr_t;
 typedef int32_t  ptrdiff_t;
-typedef uint32_t address_t;
+typedef uintptr_t address_t;
 typedef uint32_t hash_t;
 
 // kate: indent-width 4; replace-tabs on;
